qcacmn: Change data type of the code for qdf trace to uint16_t
Currently for qdf trace api data type of code is uint8_t, some of
the modules are inserting message id of type uint16_t because of
which some information is getting lost and post processing cmm
script is not able to parse these messages correctly.

To address above issue, change the data type of code for
qdf trace function.

CRs-Fixed: 2719767

+++ qdf/inc/qdf_trace.h
@@ -152,7 +152,7 @@ typedef struct qdf_trace_record_s {
 	uint64_t qtime;
 	char time[18];
 	uint8_t module;
-	uint8_t code;
+	uint16_t code;
 	uint16_t session;
 	uint32_t data;
 	uint32_t pid;
@@ -552,7 +552,7 @@ static inline void qdf_register_debugcb_init(void)
 void qdf_trace_register(QDF_MODULE_ID, tp_qdf_trace_cb);
 void qdf_trace_init(void);
 void qdf_trace_deinit(void);
-void qdf_trace(uint8_t module, uint8_t code, uint16_t session, uint32_t data);
+void qdf_trace(uint8_t module, uint16_t code, uint16_t session, uint32_t data);
 void qdf_trace_enable(uint32_t, uint8_t enable);
 void qdf_trace_dump_all(void *, uint8_t, uint8_t, uint32_t, uint32_t);
 QDF_STATUS qdf_trace_spin_lock_init(void);
@@ -574,7 +574,7 @@ void qdf_trace_enable(uint32_t bitmask_of_module_id, uint8_t enable)
 }
 
 static inline
-void qdf_trace(uint8_t module, uint8_t code, uint16_t session, uint32_t data)
+void qdf_trace(uint8_t module, uint16_t code, uint16_t session, uint32_t data)
 {
 }
